Create isEven.kt
This shows the superior density of Kotlins when-statement. The code is just 1/3 of the other implementations so it is obviously 3x better.
Unfortunately JetBrains got lazy and throws a MethodTooLargeException for an direct implementation of the algorithm. But I utilized the elegance of the type system to swipe it into a fitting solution with no performance penalties thanks to value classes.

I tested it with 3 ints and it worked, so I guess it is correct.